data: Track: chop withMainReleaseTrack
probably contains normative changes re: dropping notFoundValue
behavior in properties which inherit from main release, but has no
effect on good wiki data and maybe those compositions resolve
the same anyway...
