+++ src/data/composite/things/track/inheritFromMainRelease.js
@@ -1,10 +1,6 @@
 // Early exits with the value for the same property as specified on the
 // main release, if this track is a secondary release, and otherwise continues
 // without providing any further dependencies.
-//
-// Like withMainRelease, this will early exit (with notFoundValue) if the
-// main release is specified by reference and that reference doesn't
-// resolve to anything.
 
 import {input, templateCompositeFrom} from '#composite';
 
@@ -17,21 +13,14 @@ import withPropertyFromMainRelease
 export default templateCompositeFrom({
   annotation: `inheritFromMainRelease`,
 
-  inputs: {
-    notFoundValue: input({
-      defaultValue: null,
-    }),
-  },
-
   steps: () => [
+    raiseOutputWithoutDependency({
+      dependency: 'isSecondaryRelease',
+      mode: input.value('falsy'),
+    }),
+
     withPropertyFromMainRelease({
       property: input.thisProperty(),
-      notFoundValue: input('notFoundValue'),
-    }),
-
-    raiseOutputWithoutDependency({
-      dependency: '#isSecondaryRelease',
-      mode: input.value('falsy'),
     }),
 
     exposeDependency({
